Modify dependency constraints to install for GHC 9.2.2
I was able to install Mirage by loosening a few constraints and by
deleting the cabal.project, which I assume isn't necessary any more, as
uuagc already provides the --genmirage option?

I think it would also be cool to upload Mirage to Hackage. Would you like
to do that? I would also be happy to take care of that if it's not
convenient for you.